refactor(frontend): update person-case forms for multi-channel
