+++ frontend/app/routes/protected/multi-channel/types.d.ts
@@ -1,0 +1,11 @@
+import 'express-session';
+
+import type { PersonSinCase } from '~/.server/domain/multi-channel/services/case-api-service';
+
+declare module 'express-session' {
+  interface SessionData {
+    editingSinCase: PersonSinCase;
+  }
+}
+
+export {};
