+++ frontend/app/routes/protected/sin-application/types.d.ts
@@ -1,0 +1,114 @@
+export type BirthDetailsData = {
+  country: string;
+  province?: string;
+  city?: string;
+  fromMultipleBirth: boolean;
+};
+
+export type ContactInformationData = {
+  preferredLanguage: string;
+  primaryPhoneNumber: string;
+  secondaryPhoneNumber?: string;
+  emailAddress?: string;
+  country: string;
+  address: string;
+  postalCode: string;
+  city: string;
+  province: string;
+};
+
+export type CurrentNameData =
+  | { preferredSameAsDocumentName: true }
+  | {
+      preferredSameAsDocumentName: false;
+      firstName: string;
+      middleName?: string;
+      lastName: string;
+      supportingDocuments:
+        | { required: false }
+        | {
+            required: true;
+            documentTypes: string[];
+          };
+    };
+
+export type ParentDetailsData = (
+  | { unavailable: true }
+  | {
+      unavailable: false;
+      givenName: string;
+      lastName: string;
+      birthLocation: {
+        country: string;
+        province?: string;
+        city?: string;
+      };
+    }
+)[];
+
+export type PersonalInfoData = {
+  firstNamePreviouslyUsed?: string[];
+  lastNameAtBirth: string;
+  lastNamePreviouslyUsed?: string[];
+  gender: string;
+};
+
+export type PreviousSinData = {
+  hasPreviousSin: string;
+  socialInsuranceNumber?: string;
+};
+
+export type PrimaryDocumentData = {
+  citizenshipDate: string;
+  clientNumber: string;
+  currentStatusInCanada: string;
+  dateOfBirth: string;
+  documentType: string;
+  gender: string;
+  givenName: string;
+  lastName: string;
+  registrationNumber: string;
+};
+
+export type PrivacyStatementData = {
+  agreedToTerms: true;
+};
+
+export type RequestDetailsData = {
+  type: string;
+  scenario: string;
+};
+
+export type SecondaryDocumentData = {
+  documentType: string;
+  expiryMonth: number;
+  expiryYear: number;
+};
+
+export type InPersonSinApplication = {
+  birthDetails: BirthDetailsData;
+  contactInformation: ContactInformationData;
+  currentNameInfo: CurrentNameData;
+  parentDetails: ParentDetailsData;
+  personalInformation: PersonalInfoData;
+  previousSin: PreviousSinData;
+  primaryDocuments: PrimaryDocumentData;
+  privacyStatement: PrivacyStatementData;
+  requestDetails: RequestDetailsData;
+  secondaryDocument: SecondaryDocumentData;
+};
+
+export type Errors = Readonly<Record<string, [string, ...string[]] | undefined>> | undefined;
+
+export type LocalizedLocations = readonly Readonly<{
+  id: string;
+  alphaCode: string;
+  name: string;
+}>[];
+
+export type LocalizedOptions = readonly Readonly<{
+  id: string;
+  name: string;
+}>[];
+
+export {};
